types: Moved AudioBackend from core to types
