+++ core/src/avm2/object.rs
@@ -14,13 +14,13 @@ use crate::avm2::value::{Hint, Value};
 use crate::avm2::vector::VectorStorage;
 use crate::avm2::vtable::{ClassBoundMethod, VTable};
 use crate::avm2::Error;
-use crate::backend::audio::{SoundHandle, SoundInstanceHandle};
 use crate::bitmap::bitmap_data::BitmapData;
 use crate::context::UpdateContext;
 use crate::display_object::DisplayObject;
 use crate::html::TextFormat;
 use gc_arena::{Collect, GcCell, MutationContext};
 use ruffle_macros::enum_trait_object;
+use ruffle_types::backend::audio::{SoundHandle, SoundInstanceHandle};
 use ruffle_types::string::AvmString;
 use std::cell::{Ref, RefMut};
 use std::fmt::Debug;
