+++ core/src/character.rs
@@ -1,9 +1,9 @@
-use crate::backend::audio::SoundHandle;
 use crate::display_object::{
     Avm1Button, Avm2Button, Bitmap, EditText, Graphic, MorphShape, MovieClip, Text, Video,
 };
 use crate::font::Font;
 use gc_arena::Collect;
+use ruffle_types::backend::audio::SoundHandle;
 use ruffle_types::binary_data::BinaryData;
 
 #[derive(Clone, Collect)]
