+++ core/src/context.rs
@@ -3,10 +3,7 @@
 use crate::avm1::globals::system::SystemProperties;
 use crate::avm1::{Avm1, Object as Avm1Object, Value as Avm1Value};
 use crate::avm2::{Avm2, Object as Avm2Object, SoundChannelObject, Value as Avm2Value};
-use crate::backend::{
-    audio::{AudioBackend, AudioManager, SoundHandle, SoundInstanceHandle},
-    navigator::NavigatorBackend,
-};
+use crate::backend::{audio::AudioManager, navigator::NavigatorBackend};
 use crate::context_menu::ContextMenuState;
 use crate::display_object::{EditText, InteractiveObject, MovieClip, SoundTransform, Stage};
 use crate::external::ExternalInterface;
@@ -20,6 +17,7 @@ use core::fmt;
 use gc_arena::{Collect, MutationContext};
 use instant::Instant;
 use rand::rngs::SmallRng;
+use ruffle_types::backend::audio::{AudioBackend, SoundHandle, SoundInstanceHandle};
 use ruffle_types::backend::log::LogBackend;
 use ruffle_types::backend::render::RenderBackend;
 use ruffle_types::backend::storage::StorageBackend;
