+++ core/src/display_object/movie_clip.rs
@@ -9,8 +9,8 @@ use crate::avm2::{
     Object as Avm2Object, QName as Avm2QName, StageObject as Avm2StageObject,
     TObject as Avm2TObject, Value as Avm2Value,
 };
-use crate::backend::audio::{SoundHandle, SoundInstanceHandle};
 use bitflags::bitflags;
+use ruffle_types::backend::audio::{SoundHandle, SoundInstanceHandle};
 use ruffle_types::backend::ui::MouseCursor;
 
 use crate::avm1::activation::{Activation as Avm1Activation, ActivationIdentifier};
