+++ desktop/src/audio.rs
@@ -1,8 +1,9 @@
 use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
-use ruffle_core::backend::audio::{
-    swf, AudioBackend, AudioMixer, SoundHandle, SoundInstanceHandle, SoundTransform,
-};
+use ruffle_core::backend::audio::{swf, AudioMixer};
 use ruffle_core::impl_audio_mixer_backend;
+use ruffle_types::backend::audio::{
+    AudioBackend, SoundHandle, SoundInstanceHandle, SoundTransform,
+};
 
 pub struct CpalAudioBackend {
     #[allow(dead_code)]
