+++ types/Cargo.toml
@@ -19,6 +19,7 @@ gif = "0.11.4"
 downcast-rs = "1.2.0"
 flate2 = "1.0.24"
 generational-arena = "0.2.8"
+instant = "0.1"
 
 [dependencies.jpeg-decoder]
 version = "0.2.6"
